// src/typings/model.ts
export type AnchorConfigs = {
  /**
   * The type of anchor, with the default being a circular dot.
   * 锚点种类，默认圆点 `dot`
   */
  type?: 'dot';
  /**
   * Percentage positioning of the anchor point relative to the node.
   * 锚点相对于节点的百分比定位(0-1)
   */
  position: number[];
  /**
   * Absolute offsets of the anchor point(px).
   * 锚点相对于节点的绝对偏移(px)
   */
  offsets?: number[];
  /**
   * Absolute offsets of the end point(px).
   * 锚点连接点的绝对偏移(px)
   */
  linkOffsets?: number[];
  /**
   * Size of the anchor.
   * 锚点大小
   */
  size?: number;
  /**
   * Visibility of the anchor.
   * 锚点可见性
   */
  visible?: boolean;
  /**
   * Index of the anchor.
   * 锚点序列
   */
  index?: number;
  /**
   * The way to display the anchor.
   * 锚点展示方式
   */
  show?: 'always' | 'hover';
  /**
   * The identifier of the anchor point being snapped.
   * 锚点被吸附标识
   */
  magnet?: boolean;
  /**
   * Customize the anchor styles.
   * 自定义锚点样式
   */
  setStyles?: (nodeData: any) => Record<string, unknown>;
  /**
   * Click handler of the anchor.
   * 锚点点击回调
   */
  onClick?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse enter handler of the anchor.
   * 鼠标移入锚点回调
   */
  onMouseEnter?: (e: ShapeEvent, nodeData: any) => void;
  /**
   * Mouse leave handler of the anchor.
   * 鼠标移出锚点回调
   */
  onMouseLeave?: (e: ShapeEvent, nodeData: any) => void;
};